Dash effect works on all scenes now
Found some stuff we might wanna address:
The gameobjects for blood-splatter particle effect clones don't delete themselves after a bit, might cause performance issues if there's enough of them. To restate, the particles themselves disappear, its the gameobjects that stay.

The hit box for the sentry should be extended in the direction of the sword, right now the hit box is in the body of the sentry.

Slime collision boxes seem weird, they seem to collide with the player before the actual model of the slime comes in contact with the player

-Ryan